requirements-gpu.txt
tensorflow-gpu
numpy
gym
opencv-python
matplotlib
seaborn